Share one hash-checked bundle cache across the Remotion entry points
render-bookend reused any bundle whose index.html existed, with no hash
check at all, so it could render from a bundle that no longer matched src/
or the webpack config. render.mjs and test-render.mjs each carried their own
copy of the cache logic. One module now owns it, keyed on the compositions
and the webpack override together.
